+++ library/src/actions/filterItems/filterItems.ts
@@ -14,29 +14,34 @@ export interface FilterItemsAction<TInput extends readonly unknown[]>
    * The action reference.
    */
   readonly reference: typeof filterItems;
+  /**
+   * The filter items operation.
+   */
+  readonly operation: ArrayRequirement<TInput>;
 }
 
 /**
  * Creates a filter items transformation action.
  *
- * @param action The filter items logic.
+ * @param operation The filter items operation.
  *
  * @returns A filter items action.
  */
 export function filterItems<TInput extends readonly unknown[]>(
-  action: ArrayRequirement<TInput>
+  operation: ArrayRequirement<TInput>
 ): FilterItemsAction<TInput>;
 
 export function filterItems(
-  action: ArrayRequirement<unknown[]>
+  operation: ArrayRequirement<unknown[]>
 ): FilterItemsAction<unknown[]> {
   return {
     kind: 'transformation',
     type: 'filter_items',
     reference: filterItems,
     async: false,
+    operation,
     _run(dataset) {
-      dataset.value = dataset.value.filter(action);
+      dataset.value = dataset.value.filter(this.operation);
       return dataset;
     },
   };