+++ library/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
+- Add `operation` property to `filterItems`, `findItem`, `mapItems`, `reduceItems` and `sortItem` action
+- Rename `action` argument of `filterItems`, `findItem`, `mapItems`, `reduceItems` and `sortItem` action to `operation`
 - Rename `action` argument and property of `transform` and `transformAsync` action to `operation`
 
 ## v0.33.3 (June 19, 2024)